feat: add member delete
